+++ CHANGES.md
@@ -3,8 +3,9 @@ Version 1.14
 
 General
 -------
+* Add shortcuts links to (list of record types, current SObject RecordType and objet details) from popup  [feature 34](https://github.com/tprouvot/Chrome-Salesforce-inspector/issues/34)
 * Update manifest version from [v2](https://developer.chrome.com/docs/extensions/mv3/mv2-sunset/) to v3
-* Auto detect SObject on import page when posting data which contain SObject header [issue 30](https://github.com/tprouvot/Chrome-Salesforce-inspector/issues/30)
+* Auto detect SObject on import page when posting data which contain SObject header [feature 30](https://github.com/tprouvot/Chrome-Salesforce-inspector/issues/30)
 * Update to Salesforce API v 57.0 (Spring '23)
 * [Switch background color on import page to alert users that it's a production environnement](https://github.com/tprouvot/Chrome-Salesforce-inspector/issues/20)
 * Implement Auth2 flow to generate access token for connected App